Make abstract player controller
