+++ Assets/Prefabs/Player.prefab
@@ -45,7 +45,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d894693a13712ca42badd5cb431c9792, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape:
   - {x: 0, y: 0}
   - {x: 0, y: 1}
@@ -432,7 +432,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d894693a13712ca42badd5cb431c9792, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape:
   - {x: 0, y: 0}
   - {x: 1, y: 0}
@@ -599,7 +599,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: edf71747a000b554caa4cdd64a5e0a1e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape: []
   texture: {fileID: 2800000, guid: a4a4a539748e82b4f85898e0acff9bff, type: 3}
 --- !u!114 &1079141184
@@ -700,8 +700,6 @@ MonoBehaviour:
   xpNeededForLevel: 0a0000001e00000064000000c8000000f401000084030000
   tokensPerLevel: 020000000300000002000000030000000200000003000000
   inventory: {fileID: 1346296669}
-  levelPrefab: {fileID: 465825370906229136, guid: 11c48ec733549d04ba85c60037e738bc, type: 3}
-  repairPrefab: {fileID: 834131501170650406, guid: d154dfde0fb9d804bb2c0b657da9dae7, type: 3}
 --- !u!1 &1407947555
 GameObject:
   m_ObjectHideFlags: 0
@@ -791,7 +789,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 57f1dbebe47ad6044af02f415f18a8c2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape:
   - {x: 0, y: 0}
   - {x: 0, y: 1}
@@ -843,7 +841,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 77fd66a177722b745a09b3e546a158e3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape:
   - {x: 0, y: 0}
   - {x: 1, y: 0}
@@ -992,7 +990,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7213755287841571539}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8da0605e03851884b8ef65530fb6c913, type: 3}
+  m_Script: {fileID: 11500000, guid: 9839d6b838ad2434dbac8d959ee4cd59, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   QueueMovesOnButton: 1
@@ -1016,7 +1014,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 724217e690c64c249b57d39309f9db26, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
 --- !u!114 &991764315
 MonoBehaviour:
   m_ObjectHideFlags: 0
