+++ Assets/Scripts/Lootables/Uplink.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 public class Uplink : Lootable
 {
