+++ Assets/Scripts/Primitives/GriddedEntity.cs
@@ -1,45 +1,6 @@
-using UnityEngine;
-using DeCrawl.Primitives;
+using DeCrawl.World;
 
 /// <summary>
 /// Utility class for things that exists on the level grid
 /// </summary>
-[ExecuteInEditMode]
-public class GriddedEntity : MonoBehaviour
-{
-    [SerializeField]
-    Vector3Int gridPosition;
-
-    private void Awake()
-    {
-        var movable = GetComponent<MovingEntity>();
-        if (movable) { 
-            movable.OnMove += GriddedEntity_OnMove;
-        }
-    }
-
-    private void OnDestroy()
-    {
-        var movable = GetComponent<MovingEntity>();
-        if (movable)
-        {
-            movable.OnMove -= GriddedEntity_OnMove;
-        }
-    }
-
-    private void GriddedEntity_OnMove(string id, Vector3Int position, CardinalDirection lookDirection)
-    {
-        gridPosition = position;
-    }
-
-
-    private void Update()
-    {
-        if (!Application.isPlaying)
-        {
-            transform.position = Level.instance.AsWorldPosition(gridPosition);
-        }
-    }
-
-
-}
+public class GriddedEntity : AbstractGriddedEntity<GridEntity, bool> {}
