+++ Assets/deCrawl/Scrips/World/AbstractGriddedEntity.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9839d6b838ad2434dbac8d959ee4cd59
+guid: 5383707dca48de34badb72a9e95d0742
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
