+++ Assets/deCrawl/Scrips/World/AbstractPlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9839d6b838ad2434dbac8d959ee4cd59
+guid: 8da0605e03851884b8ef65530fb6c913
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
